Staging

+++ app/controllers/backoffice/raffles_carts_controller.rb
@@ -7,10 +7,12 @@ module Backoffice
     end
 
     def destroy
-      order_item = current_raffles_cart.order_items.find(params[:order_item_id])
-      order_item.raffle_ticket.update!(order_item_id: nil, status: :available, user_id: nil)
-      order_item.destroy
-      Shopping::UpdateCartTotals.call(current_raffles_cart, params[:country])
+      order_item = current_raffles_cart.order_items.where(id: params[:order_item_id]).first
+      if order_item.present?
+        order_item.raffle_ticket.update!(order_item_id: nil, status: :available, user_id: nil)
+        order_item.destroy
+        Shopping::UpdateCartTotals.call(current_raffles_cart, params[:country])
+      end
       flash[:success] = t(:removed_successfully)
 
       redirect_to backoffice_raffles_carts_path
